+++ eras/shelley/impl/CHANGELOG.md
@@ -28,7 +28,8 @@
   * `scriptCred` in favor of `credScriptHash`
   * `scriptStakeCred` in favor of `getScriptWitnessTxCert`
   * `requiresVKeyWitness` in favor of `getVKeyWitnessTxCert`
-  * `delegCWitness`
+  * `delegCWitness` - no longer used.
+  * `propWits` - will become an internal function in the future version
 * `validateNeededWitnesses` no longer accepts `witsVKeyNeeded` as an argument.
 
 ## 1.2.0.0